+++ src/core/network/UDPSocket.hpp
@@ -20,8 +20,8 @@ public:
     ~UDPSocket();
     
     int bindSocket(const SocketAddress& toAddress);
-    int sendToAddress(const void* toSend, int length, const SocketAddress* toAddress);
-    int receiveFromAddress(void* toReceive, int maxLength, SocketAddress& fromAddress);
+    int sendToAddress(const void* toSend, uint16_t length, const SocketAddress* toAddress);
+    int receiveFromAddress(void* toReceive, uint16_t maxLength, SocketAddress& fromAddress);
     int setNonBlockingMode(bool nonBlocking);
     
 private:
